requirements.txt
streamlit>=1.35.0
groq>=0.8.0
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
